+++ CAN.Webwinkel/src/CAN.Webwinkel/Startup.cs
@@ -89,7 +89,7 @@ namespace CAN.Webwinkel
             //o => new OnderhoudsServiceAgent() { BaseUri = new Uri("http://lapiwe-onderhoudservice:80") }
             services.AddDbContext<WinkelDatabaseContext>(options => options.UseSqlServer(Environment.GetEnvironmentVariable("dbconnectionstring")));
             services.AddScoped<IRepository<Categorie, int>, CategorieRepository>();
-            services.AddScoped<IRepository<Artikel, int>, ArtikelRepository>();
+            services.AddScoped<IRepository<Artikel, long>, ArtikelRepository>();
             services.AddScoped<ICategorieService, CategorieService>();
             services.AddScoped<IArtikelService, ArtikelService>();
             services.AddScoped<IKlantAgent, KlantAgent>(s => new KlantAgent() { BaseUri = new Uri("http://can-klantbeheer:80") });
